+++ front/app/modules/commercial/customizable_navbar/admin/containers/HiddenNavbarItemList/HiddenNavbarItemList.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { render, screen, fireEvent } from 'utils/testUtils/rtl';
-import HiddenNavbarItemList from './HiddenNavbarItemList';
+import HiddenNavbarItemList from '.';
 import allNavbarItems from 'hooks/fixtures/navbarItems';
-import { addNavbarItem } from '../../services/navbar';
-import { deletePage } from 'services/pages';
+import { addNavbarItem } from '../../../services/navbar';
+import { deletePage } from '../../../services/pages';
 
 jest.mock('services/locale');
 jest.mock('services/appConfiguration');
@@ -12,7 +12,7 @@ let mockNavbarItems = allNavbarItems;
 const mockRemovedDefaultNavbarItems = [];
 
 jest.mock('hooks/useNavbarItems', () => jest.fn(() => mockNavbarItems));
-jest.mock('../../hooks/useRemovedDefaultNavbarItems', () =>
+jest.mock('../../../hooks/useRemovedDefaultNavbarItems', () =>
   jest.fn(() => mockRemovedDefaultNavbarItems)
 );
 
@@ -20,12 +20,12 @@ jest.mock('hooks/usePages');
 jest.mock('hooks/usePageSlugById');
 jest.mock('hooks/useLocale');
 
-jest.mock('../../services/navbar', () => ({
+jest.mock('../../../services/navbar', () => ({
   addNavbarItem: jest.fn(),
 }));
 
-jest.mock('services/pages', () => {
-  const original = jest.requireActual('services/pages');
+jest.mock('../../../services/pages', () => {
+  const original = jest.requireActual('../../../services/pages');
 
   return {
     ...original,