+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
-## Next
+## 2021-12-23
 
+### Added
+
+- Adding pages in 'Navigation' tab in settings now possible, changing names of navbar items now works, removed 'secret pages-page'.
 
 ## 2021-12-22